Do not instantiate MarkdownFormatHelpers
No need: it's just static methods; also instead of passing references,
just import as needed.
